+++ server/routes/index.js
@@ -2,12 +2,12 @@ const path = require('path');
 const keystone = require('keystone');
 const cors = require('cors');
 
-const Depoimentos = keystone.list('Depoimentos');
-const Servicos = keystone.list('Servicos');
-const SobreNos = keystone.list('SobreNos');
-const Valores = keystone.list('Valores');
-const Banner = keystone.list('Banner');
-const Footer = keystone.list('Footer');
+const depoimentos_controller = require('../controllers/depoimentosController');
+const servicos_controller = require('../controllers/servicosController');
+const sobreNos_controller = require('../controllers/sobreNosController');
+const valores_controller = require('../controllers/valoresController');
+const banner_controller = require('../controllers/bannerController');
+const footer_controller = require('../controllers/footerController');
 
 module.exports = (app) => {
   app.use(cors());
@@ -16,65 +16,12 @@ module.exports = (app) => {
     res.sendFile(path.join(__dirname, '../public/index.html'));
   });
 
-  app.get('/api/depoimentos', (req, res) => {
-    Depoimentos.model.find((err, data) => {
-      if (err) {
-        res.status(500).send('DB Error');
-      } else {
-        res.send(data);
-      }
-    });
-  });
-
-  app.get('/api/servicos', (req, res) => {
-    Servicos.model.find((err, data) => {
-      if (err) {
-        res.status(500).send('DB Error');
-      } else {
-        res.send(data);
-      }
-    });
-  });
-
-  app.get('/api/banner', (req, res) => {
-    Banner.model.find((err, data) => {
-      if (err) {
-        res.status(500).send('DB Error');
-      } else {
-        res.send(data);
-      }
-    });
-  });
-  
-  app.get('/api/sobrenos', (req, res) => {
-    SobreNos.model.find((err, data) => {
-      if (err) {
-        res.status(500).send(('DB Error'));
-      } else {
-        res.send(data);
-      }
-    });
-  });
-
-  app.get('/api/valores', (req, res) => {
-    Valores.model.find((err, data) => {
-      if (err) {
-        res.status(500).send(('DB Error'));
-      } else {
-        res.send(data);
-      }
-    });
-  });
-
-  app.get('/api/footer', (req, res) => {
-    Footer.model.find((err, data) => {
-      if (err) {
-        res.status(500).send(('DB Error'));
-      } else {
-        res.send(data);
-      }
-    });
-  });
+  app.get('/api/depoimentos', depoimentos_controller.getData);
+  app.get('/api/servicos', servicos_controller.getData);
+  app.get('/api/sobrenos', sobreNos_controller.getData);
+  app.get('/api/valores', valores_controller.getData);
+  app.get('/api/banner', banner_controller.getData);
+  app.get('/api/footer', footer_controller.getData);
 
   app.get('*', (req, res) => {
 		res.redirect('/');
